[MIG] document_quick_access: Migration to 17.0
