+++ document_quick_access/static/src/xml/document_quick_access_launcher.xml
@@ -1,18 +1,15 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <templates>
     <t t-name="document_quick_access_launcher.view.Menu">
-        <div class="o_mail_systray_item dropdown">
-            <a
-                class="dropdown-toggle o-no-caret"
-                data-display="static"
-                role="button"
-                aria-expanded="false"
+        <div>
+            <button
+                class="o_nav_entry"
                 title="QR search"
-                href="#"
                 accesskey="q"
+                t-on-click="onClick"
             >
                 <i class="fa fa-qrcode" />
-            </a>
+            </button>
         </div>
     </t>
 </templates>
